DPAD-1089 VIDEO-229 :: Moved types out of GetYoutubeTakeoverDetails and into the youtube/types.d.ts file
